+++ scritpt_Sequencial.py
@@ -18,7 +18,7 @@ def salvar_resultados_em_arquivo(valores_T, resultados, tempo_medio, nome_arquiv
 
 
 inicio_T = 0
-fim_T = 400
+fim_T = 500
 passo_T = 100
 
 valores_T = list(range(inicio_T, fim_T + 1, passo_T))
@@ -41,7 +41,9 @@ for T in valores_T:
         print(f"Erro ao executar o programa com T={T}: {resultado.stderr}")
     else:
         try:
-            tempo_execucao = float(resultado.stdout.strip())
+            # Filtrar a saída para obter apenas a última linha que contém o tempo
+            linhas_saida = resultado.stdout.strip().split("\n")
+            tempo_execucao = float(linhas_saida[-1])  # Última linha contém o tempo
             resultados.append(tempo_execucao)
             print(f"Execução concluída para T={T}. Tempo: {tempo_execucao} segundos")
         except ValueError:
@@ -50,14 +52,16 @@ for T in valores_T:
 tempo_medio = tempo_medio(resultados)
 print(f"Tempo médio: {tempo_medio}")
 
-plt.plot(valores_T, resultados, marker='o')
-plt.title("Análise de Tempo de Execução")
-plt.xlabel("Número de interações")
-plt.ylabel("Tempo de Execução (segundos)")
-plt.grid(True)
-plt.savefig("analise_tempo_sequencial.png")
-# plt.show()
-
-salvar_resultados_em_arquivo(valores_T, resultados, tempo_medio)
+if resultados:  # Garantir que existem resultados para plotar
+    plt.plot(valores_T[:len(resultados)], resultados, marker='o')
+    plt.title("Análise de Tempo de Execução")
+    plt.xlabel("Número de interações")
+    plt.ylabel("Tempo de Execução (segundos)")
+    plt.grid(True)
+    plt.savefig("analise_tempo_sequencial.png")
+    # plt.show()
+    salvar_resultados_em_arquivo(valores_T[:len(resultados)], resultados, tempo_medio)
+else:
+    print("Nenhum resultado válido foi obtido para plotar o gráfico.")
 
 print("Todas as execuções foram concluídas.")
